+++ test/new_relic/agent/agent/start_test.rb
@@ -5,7 +5,8 @@
 require File.expand_path(File.join(File.dirname(__FILE__), '..', '..', '..', 'test_helper'))
 class NewRelic::Agent::Agent::StartTest < Minitest::Test
   require 'new_relic/agent/agent'
-  include NewRelic::Agent::Agent::Start
+  include NewRelic::Agent::Startup
+  include NewRelic::Agent::SpecialStartup
 
   def setup
     @harvester = stub("dummy harvester")
